Speed-up the whole module
It was originally anticipated that moving data tables from P2001.py to
P2001.npz would speed up the whole module, however this was not
completely the case because np.load() actually creates an object where
the arrays are mapped in memory with lazy access and unzip/deflate
on-the-fly. As a consequence, speed was still sub-optimal.

This commit unpacks the arrays in memory, making all further access faster.